articulos/requirements.txt
Django>=3.2,<4.0
mysqlclient
drf-yasg
djangorestframework
djangorestframework-simplejwt
django-environ